chore(Maybe): Upgrade the typing for maybe.sequence
We would like to have a better typing of sequence such that when we use sequence we get returned a typing with same informational entropy as what was inputed.
Example type difference
```
from: typeof sequence({ a1: Monad<A>, a2: Monad<A> }) === Monad<{ [string]: A }>
to: typeof sequence({ a1: Monad<A>, a2: Monad<A> }) === Monad<{ a1: A, a2: A}>
```
and
```
from: typeof sequence<A | B>({ a: Monad<A>, b: Monad<B> }) === Monad<{ [string]: A | B }>
to: typeof sequence({ a: Monad<A>, b: Monad<B> }) === Monad<{ a: A, b: B}>
```

Breaking: Previous typings that specified the generic for funciton of all or sequence will fail.